+++ lib/model/settings/modules/app.dart
@@ -1,22 +1,12 @@
 import 'dart:ui' as ui;
 
-import 'package:aves/model/filters/covered/tag.dart';
-import 'package:aves/model/filters/filters.dart';
 import 'package:aves/model/settings/defaults.dart';
-import 'package:aves/model/source/collection_source.dart';
-import 'package:aves/model/vaults/vaults.dart';
 import 'package:aves/utils/calendar/aves_locale.dart';
 import 'package:aves/widgets/aves_app.dart';
 import 'package:aves_model/aves_model.dart';
 import 'package:flutter/widgets.dart';
 
 mixin AppSettings on SettingsAccess {
-  static const int recentFilterHistoryMax = 20;
-
-  void initAppSettings() {
-    vaults.lockStateChangeNotifier.addListener(_onVaultsChanged);
-  }
-
   bool get hasAcceptedTerms => getBool(SettingKeys.hasAcceptedTermsKey) ?? SettingsDefaults.hasAcceptedTerms;
 
   set hasAcceptedTerms(bool newValue) => set(SettingKeys.hasAcceptedTermsKey, newValue);
@@ -146,35 +136,4 @@ mixin AppSettings on SettingsAccess {
   List<int>? get topEntryIds => getStringList(SettingKeys.topEntryIdsKey)?.map(int.tryParse).nonNulls.toList();
 
   set topEntryIds(List<int>? newValue) => set(SettingKeys.topEntryIdsKey, newValue?.map((id) => id.toString()).nonNulls.toList());
-
-  List<String> get recentDestinationAlbums => getStringList(SettingKeys.recentDestinationAlbumsKey) ?? [];
-
-  set recentDestinationAlbums(List<String> newValue) => set(SettingKeys.recentDestinationAlbumsKey, newValue.take(recentFilterHistoryMax).toList());
-
-  // recent tags
-
-  List<CollectionFilter> get _recentTags => (getStringList(SettingKeys.recentTagsKey) ?? []).map(CollectionFilter.fromJson).nonNulls.toList();
-
-  set _recentTags(List<CollectionFilter> newValue) => set(SettingKeys.recentTagsKey, newValue.take(recentFilterHistoryMax).map((filter) => filter.toJsonString()).toList());
-
-  // when vaults are unlocked, recent tags are transient and not persisted
-  List<CollectionFilter>? _protectedRecentTags;
-
-  List<CollectionFilter> get recentTags => vaults.needProtection ? _protectedRecentTags ?? List.of(_recentTags) : _recentTags;
-
-  set recentTags(List<CollectionFilter> newValue) {
-    if (vaults.needProtection) {
-      _protectedRecentTags = newValue;
-    } else {
-      _recentTags = newValue;
-    }
-  }
-
-  void _onVaultsChanged() => _protectedRecentTags = null;
-
-  void removeObsoleteRecentTags(CollectionSource? source) {
-    if (source != null) {
-      recentTags = recentTags.where((v) => v is! TagFilter || source.sortedTags.contains(v.tag)).toList();
-    }
-  }
 }