+++ lib/widgets/debug/settings.dart
@@ -1,6 +1,7 @@
 import 'package:aves/model/settings/settings.dart';
 import 'package:aves/widgets/collection/collection_page.dart';
 import 'package:aves/widgets/common/identity/aves_expansion_tile.dart';
+import 'package:aves/widgets/common/identity/highlight_title.dart';
 import 'package:aves/widgets/filter_grids/albums_page.dart';
 import 'package:aves/widgets/filter_grids/countries_page.dart';
 import 'package:aves/widgets/filter_grids/places_page.dart';
@@ -52,37 +53,104 @@ class _DebugSettingsSectionState extends State<DebugSettingsSection> with Automa
               title: const Text('canUseAnalysisService'),
             ),
             Padding(
-              padding: const EdgeInsets.only(left: 8, right: 8, bottom: 8),
+              padding: const EdgeInsets.all(8),
               child: InfoRowGroup(
                 info: {
                   'catalogTimeZoneRawOffsetMillis': '${settings.catalogTimeZoneOffsetMillis}',
-                  'tileExtent - Collection': '${settings.getTileExtent(CollectionPage.routeName)}',
-                  'tileExtent - Albums': '${settings.getTileExtent(AlbumListPage.routeName)}',
-                  'tileExtent - Countries': '${settings.getTileExtent(CountryListPage.routeName)}',
-                  'tileExtent - Places': '${settings.getTileExtent(PlaceListPage.routeName)}',
-                  'tileExtent - Tags': '${settings.getTileExtent(TagListPage.routeName)}',
-                  'mapStyle': '${settings.mapStyle}',
-                  'infoMapZoom': '${settings.infoMapZoom}',
-                  'customMapStyles': toMultiline(settings.customMapStyles),
                   'collectionSelectionQuickActions': '${settings.collectionSelectionQuickActions}',
                   'viewerQuickActions': '${settings.viewerQuickActions}',
-                  'drawerTypeBookmarks': toMultiline(settings.drawerTypeBookmarks),
-                  'drawerAlbumBookmarks': toMultiline(settings.drawerAlbumBookmarks),
-                  'drawerPageBookmarks': toMultiline(settings.drawerPageBookmarks),
                   'pinnedFilters': toMultiline(settings.pinnedFilters),
                   'hiddenFilters': toMultiline(settings.hiddenFilters),
                   'deactivatedHiddenFilters': toMultiline(settings.deactivatedHiddenFilters),
+                  'topEntryIds': '${settings.topEntryIds}',
+                  'longPressTimeout': '${settings.longPressTimeout}',
+                },
+              ),
+            ),
+            const Padding(
+              padding: EdgeInsets.all(8),
+              child: HighlightTitle(title: 'Drawer'),
+            ),
+            Padding(
+              padding: const EdgeInsets.all(8),
+              child: InfoRowGroup(
+                info: {
+                  'drawerTypeBookmarks': toMultiline(settings.drawerTypeBookmarks),
+                  'drawerAlbumBookmarks': toMultiline(settings.drawerAlbumBookmarks),
+                  'drawerPageBookmarks': toMultiline(settings.drawerPageBookmarks),
+                },
+              ),
+            ),
+            const Padding(
+              padding: EdgeInsets.all(8),
+              child: HighlightTitle(title: 'Groups'),
+            ),
+            Padding(
+              padding: const EdgeInsets.all(8),
+              child: InfoRowGroup(
+                info: {
+                  'albumGroups': toMultiline(settings.albumGroups.entries),
+                  'tagGroups': toMultiline(settings.tagGroups.entries),
+                },
+              ),
+            ),
+            const Padding(
+              padding: EdgeInsets.all(8),
+              child: HighlightTitle(title: 'History'),
+            ),
+            Padding(
+              padding: const EdgeInsets.all(8),
+              child: InfoRowGroup(
+                info: {
+                  'recentSettingKeys': toMultiline(settings.recentSettingKeys),
                   'searchHistory': toMultiline(settings.searchHistory),
                   'recentDestinationAlbums': toMultiline(settings.recentDestinationAlbums),
                   'recentTags': toMultiline(settings.recentTags),
-                  'locale (basic)': '${settings.basicLocale}',
-                  'locale (resolved)': '${settings.resolvedLocale}',
-                  'locale (Aves)': '${settings.avesLocale}',
-                  'systemLocales': '${WidgetsBinding.instance.platformDispatcher.locales}',
-                  'topEntryIds': '${settings.topEntryIds}',
-                  'longPressTimeout': '${settings.longPressTimeout}',
-                  'albumGroups': toMultiline(settings.albumGroups.entries),
-                  'tagGroups': toMultiline(settings.tagGroups.entries),
+                },
+              ),
+            ),
+            const Padding(
+              padding: EdgeInsets.all(8),
+              child: HighlightTitle(title: 'Locale'),
+            ),
+            Padding(
+              padding: const EdgeInsets.all(8),
+              child: InfoRowGroup(
+                info: {
+                  'basic': '${settings.basicLocale}',
+                  'resolved': '${settings.resolvedLocale}',
+                  'aves': '${settings.avesLocale}',
+                  'system': '${WidgetsBinding.instance.platformDispatcher.locales}',
+                },
+              ),
+            ),
+            const Padding(
+              padding: EdgeInsets.all(8),
+              child: HighlightTitle(title: 'Map'),
+            ),
+            Padding(
+              padding: const EdgeInsets.all(8),
+              child: InfoRowGroup(
+                info: {
+                  'mapStyle': '${settings.mapStyle}',
+                  'infoMapZoom': '${settings.infoMapZoom}',
+                  'customMapStyles': toMultiline(settings.customMapStyles),
+                },
+              ),
+            ),
+            const Padding(
+              padding: EdgeInsets.all(8),
+              child: HighlightTitle(title: 'Tile Extent'),
+            ),
+            Padding(
+              padding: const EdgeInsets.all(8),
+              child: InfoRowGroup(
+                info: {
+                  'collection': '${settings.getTileExtent(CollectionPage.routeName)}',
+                  'albums': '${settings.getTileExtent(AlbumListPage.routeName)}',
+                  'countries': '${settings.getTileExtent(CountryListPage.routeName)}',
+                  'places': '${settings.getTileExtent(PlaceListPage.routeName)}',
+                  'tags': '${settings.getTileExtent(TagListPage.routeName)}',
                 },
               ),
             ),