+++ lib/widgets/settings/language/locale_tile.dart
@@ -5,12 +5,15 @@ import 'package:aves/theme/durations.dart';
 import 'package:aves/widgets/common/extensions/build_context.dart';
 import 'package:aves/widgets/settings/language/locale_selection_page.dart';
 import 'package:aves/widgets/settings/language/locales.dart';
+import 'package:aves_model/aves_model.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:provider/provider.dart';
 
 class LocaleTile extends StatelessWidget {
   static const systemLocaleOption = ui.Locale('system');
+
+  static const List<String> settingKeys = [SettingKeys.localeKey];
 
   const LocaleTile({super.key});