+++ lib/widgets/settings/settings_search_delegate.dart
@@ -1,3 +1,4 @@
+import 'package:aves/model/settings/settings.dart';
 import 'package:aves/theme/icons.dart';
 import 'package:aves/widgets/common/extensions/build_context.dart';
 import 'package:aves/widgets/common/identity/empty.dart';
@@ -6,6 +7,8 @@ import 'package:aves/widgets/common/providers/media_query_data_provider.dart';
 import 'package:aves/widgets/common/search/delegate.dart';
 import 'package:aves/widgets/settings/settings_definition.dart';
 import 'package:flutter/material.dart';
+
+typedef _SectionPredicate = Future<List<SettingsTile>> Function(BuildContext context, SettingsSection section);
 
 class SettingsSearchDelegate extends AvesSearchDelegate {
   final List<SettingsSection> sections;
@@ -23,14 +26,26 @@ class SettingsSearchDelegate extends AvesSearchDelegate {
   @override
   Widget buildSuggestions(BuildContext context) {
     final upQuery = query.toUpperCase().trim();
-    if (upQuery.isEmpty) return const SizedBox();
-
-    bool testKey(String key) => key.toUpperCase().contains(upQuery);
+    _SectionPredicate testSection;
+    if (query.trim().isEmpty) {
+      final recentSettingKeys = settings.recentSettingKeys;
+      testSection = (context, section) async {
+        final allTiles = await section.tiles(context);
+        final filteredTiles = allTiles.where((v) => v.settingKeys.any(recentSettingKeys.contains)).toList();
+        return filteredTiles;
+      };
+    } else {
+      bool testTitle(String key) => key.toUpperCase().contains(upQuery);
+      testSection = (context, section) async {
+        final allTiles = await section.tiles(context);
+        final filteredTiles = testTitle(section.title(context)) ? allTiles : allTiles.where((v) => testTitle(v.title(context))).toList();
+        return filteredTiles;
+      };
+    }
 
     final loader = Future.wait(
       sections.map((section) async {
-        final allTiles = await section.tiles(context);
-        final filteredTiles = testKey(section.title(context)) ? allTiles : allTiles.where((v) => testKey(v.title(context))).toList();
+        List<SettingsTile> filteredTiles = await testSection(context, section);
         if (filteredTiles.isEmpty) return null;
 
         return (context) {