+++ plugins/aves_model/lib/src/settings/keys.dart
@@ -4,12 +4,15 @@ class SettingKeys {
   static const Set<String> _internalKeys = {
     hasAcceptedTermsKey,
     catalogTimeZoneOffsetMillisKey,
-    searchHistoryKey,
     platformAccelerometerRotationKey,
     platformTransitionAnimationScaleKey,
     topEntryIdsKey,
+    // history
+    recentSettingKeysKey,
+    searchHistoryKey,
     recentDestinationAlbumsKey,
     recentTagsKey,
+    // debug
     debugShowViewerTilesKey,
   };
 
@@ -29,6 +32,11 @@ class SettingKeys {
   static const tileLayoutPrefixKey = 'tile_layout_';
   static const entryRenamingPatternKey = 'entry_renaming_pattern';
   static const topEntryIdsKey = 'top_entry_ids';
+
+  // history
+  static const saveSearchHistoryKey = 'save_search_history';
+  static const searchHistoryKey = 'search_history';
+  static const recentSettingKeysKey = 'recent_setting_keys';
   static const recentDestinationAlbumsKey = 'recent_destination_albums';
   static const recentTagsKey = 'recent_tags';
 
@@ -153,10 +161,6 @@ class SettingKeys {
   static const mapDefaultCenterKey = 'map_default_center';
   static const mapShowItemTracksKey = 'map_show_item_tracks';
   static const customMapStylesKey = 'custom_map_styles';
-
-  // search
-  static const saveSearchHistoryKey = 'save_search_history';
-  static const searchHistoryKey = 'search_history';
 
   // bin
   static const enableBinKey = 'enable_bin';